registro medianamente funcional ya agrega al usuario y todo el rollo pero tenemos varios problemas, el primero es que como no creo que el usuario sepa su UID hasta que se le asigne su tarjeta ps no se lo pedimos en su registro y su UID es su PK entonces no se pueder repetir asi que yo creo que vamos a tener que meter una funcion de randomizer o algo asi para que jale, la segunda problematica es q no se que retornar o q para que cuando se realice el registro te llegue retroaliementacion del pieichpi y asi se vea mas perro, si corres el programa lo vas a entender, y aparte nomas le falta diseño y se puede decir que esta funcional
