+++ app/src/main/java/com/example/gymbud/MainActivity.java
@@ -62,41 +62,6 @@ SharedPreferences archivo;
 
     }
 
-    public void LinkRegistrarse(View view) {
-        String url = "http://francoaldrete.com/GymBud/bd.php?usr=";
-        url = url + ETusr.getText().toString();
-        url = url + "&pass=";
-        url = url + ETcontra.getText().toString();
-        RequestQueue lanzarPeticion= Volley.newRequestQueue(this);
-        JsonObjectRequest pet = new JsonObjectRequest(Request.Method.GET, url, null,
-                new Response.Listener<JSONObject>() {
-            @Override
-            public void onResponse(JSONObject response) {
-                String UID,user,mail,password,presence;
-                try {
-                    UID = response.getString("UID");
-                    user = response.getString("User");
-                    mail = response.getString("Mail");
-                    password = response.getString("Password");
-                    presence = response.getString("Presence");
-                        Intent i = new Intent(MainActivity.this, infopersonal.class);
-                        startActivity(i);
-                        finish();
-
-                } catch (JSONException e) {
-                    e.printStackTrace();
-                }
-                Toast.makeText(MainActivity.this,response.toString(),Toast.LENGTH_SHORT).show();
-            }
-        }, new Response.ErrorListener() {
-            @Override
-            public void onErrorResponse(VolleyError error) {
-                Log.d("yo", error.getMessage());
-            }
-        });
-        lanzarPeticion.add(pet);
-    }
-
     public void clickInicio(View view) {
 
         //Toast.makeText(this, "picon",Toast.LENGTH_SHORT).show();
