+++ app/src/main/res/layout/activity_registro.xml
@@ -4,6 +4,118 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:background="#252525"
     tools:context=".Registro">
 
+    <EditText
+        android:id="@+id/ETcorreo"
+        android:layout_width="324dp"
+        android:layout_height="58dp"
+        android:autofillHints="name"
+        android:backgroundTint="@color/white"
+        android:fontFamily="sans-serif-condensed"
+        android:gravity="center"
+        android:hint="Correo"
+        android:singleLine="true"
+        android:text=""
+        android:textColor="@color/white"
+        android:textColorHint="@color/white"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.494"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="0.433" />
+
+    <TextView
+        android:id="@+id/textView"
+        android:layout_width="324dp"
+        android:layout_height="58dp"
+        android:autofillHints="name"
+        android:backgroundTint="@color/white"
+        android:fontFamily="sans-serif-condensed"
+        android:gravity="center"
+        android:singleLine="true"
+        android:text="REGISTRO"
+        android:textColor="@color/white"
+        android:textColorHint="@color/white"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.498"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="0.046" />
+
+    <EditText
+        android:id="@+id/ETcontrasenaconf"
+        android:layout_width="324dp"
+        android:layout_height="58dp"
+        android:autofillHints="name"
+        android:backgroundTint="@color/white"
+        android:fontFamily="sans-serif-condensed"
+        android:gravity="center"
+        android:hint="Confirmar contraseña"
+        android:singleLine="true"
+        android:text=""
+        android:textColor="@color/white"
+        android:textColorHint="@color/white"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.494"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="0.739" />
+
+    <EditText
+        android:id="@+id/ETusuario"
+        android:layout_width="324dp"
+        android:layout_height="58dp"
+        android:autofillHints="name"
+        android:backgroundTint="@color/white"
+        android:fontFamily="sans-serif-condensed"
+        android:gravity="center"
+        android:hint="@string/usuario"
+        android:singleLine="true"
+        android:text=""
+        android:textColor="@color/white"
+        android:textColorHint="@color/white"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.494"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="0.279" />
+
+    <EditText
+        android:id="@+id/ETcontrasena"
+        android:layout_width="324dp"
+        android:layout_height="58dp"
+        android:autofillHints="name"
+        android:backgroundTint="@color/white"
+        android:fontFamily="sans-serif-condensed"
+        android:gravity="center"
+        android:hint="Contraseña"
+        android:singleLine="true"
+        android:text=""
+        android:textColor="@color/white"
+        android:textColorHint="@color/white"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.494"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="0.595" />
+
+    <Button
+        android:id="@+id/Bconfirmar"
+        android:layout_width="184dp"
+        android:layout_height="65dp"
+        android:text="Confirmar"
+        android:onClick="confirmar"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.497"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="0.896" />
 </androidx.constraintlayout.widget.ConstraintLayout>
